\name{algorithm}
\alias{algorithm}
\alias{algorithm-methods}
\alias{algorithm,baselineAlgTest-method}
\alias{extraArgs}
\alias{extraArgs-methods}
\alias{extraArgs,baselineAlgTest-method}
\title{
  Extraction methods for "baselineAlgTest" objects
}
\description{
  Extraction methods specifically for objects of class \code{\linkS4class{baselineAlgTest}}
}
\usage{
algorithm(object)
extraArgs(object)
}
\arguments{
  \item{object}{Object of class \code{\linkS4class{baselineAlgTest}}}
}
\value{
  The corresponding slot
}
\author{Bjørn-Helge Mevik and Kristian Hovde Liland}
\seealso{
  \code{\linkS4class{baselineAlgTest}}
}
\keyword{baseline}
\keyword{spectra}
