\name{funcName}
\alias{funcName}
\alias{funcName-methods}
\alias{funcName,baselineAlg-method}
\alias{funcName,baselineAlgTest-method}
\title{
  Extract the "funcName" slot.
}
\description{
  Extract the \code{funcName} slot from an object of class
  \code{\linkS4class{baselineAlg}} or \code{\linkS4class{baselineAlgTest}}
}
\usage{
funcName(object)
}
\arguments{
  \item{object}{An object of class
  \code{\linkS4class{baselineAlg}} or \code{\linkS4class{baselineAlgTest}}}
}
\value{
  The \code{funcName} slot of the object.
}
\author{Bjørn-Helge Mevik and Kristian Hovde Liland}
\seealso{
  \code{\linkS4class{baselineAlg}}, \code{\linkS4class{baselineAlgTest}}
}
\keyword{spectra}
\keyword{baseline}
