%%% $Id: ind.min.Rd 193 2012-06-24 21:13:42Z kristl $
\name{ind.min}
\alias{ind.min}
\alias{ind.min-methods}
\alias{ind.min,predictionResult-method}
\alias{paramName}
\alias{paramName-methods}
\alias{paramName,predictionResult-method}
\title{Extraction methods specific for "predictionResult" objects}
\description{
  Extract information from objects of class
  \code{\linkS4class{predictionResult}}.
}
\usage{
ind.min(object)
paramName(object)
}
\arguments{
  \item{object}{Object of class \code{\linkS4class{predictionResult}}}
}
\value{
  The corresponding slot of the object.
}
\author{Bjørn-Helge Mevik and Kristian Hovde Liland}
\seealso{
  \code{\linkS4class{predictionResult}}
}
\keyword{methods}
