%%% $Id: param.ind.min.Rd 193 2012-06-24 21:13:42Z kristl $
\name{param.ind.min}
\alias{param.ind.min}
\alias{param.ind.min-methods}
\alias{param.ind.min,baselineAlgResult-method}
\alias{qualMeas.ind.min}
\alias{qualMeas.ind.min-methods}
\alias{qualMeas.ind.min,baselineAlgResult-method}
\title{
  Extraction methods for "baselineAlgResult" objects
}
\description{
  Extraction methods that are specific for objects of class
  \code{\linkS4class{baselineAlgResult}}
}
\usage{
param.ind.min(object)
qualMeas.ind.min(object)
}
\arguments{
  \item{object}{Object of class \code{\linkS4class{baselineAlgResult}}}
}
\value{
  The corresponding slot
}
\author{Bjørn-Helge Mevik and Kristian Hovde Liland}
\seealso{
  Class \code{\linkS4class{baselineAlgResult}}
}
\keyword{baseline}
\keyword{spectra}
