%%% $Id: name.Rd 193 2012-06-24 21:13:42Z kristl $
\name{name}
\alias{name}
\alias{name-methods}
\alias{name,baselineAlg-method}
\alias{description}
\alias{description-methods}
\alias{description,baselineAlg-method}
\title{Extraction methods for "baselineAlg" objects}
\description{Extraction methods specifically for objects of class \code{\linkS4class{baselineAlg}}}
\usage{
name(object)
description(object)
}
\arguments{
  \item{object}{Object of class \code{\linkS4class{baselineAlg}}}
}
\value{
  The methods return the corresponding slot of the object.
}
\author{Bjørn-Helge Mevik and Kristian Hovde Liland}
\seealso{
  \code{\linkS4class{baselineAlg}}, \code{\link{funcName}}.
}
\keyword{baseline}
\keyword{spectra}
