%%% $Id: predictionTest-class.Rd 193 2012-06-24 21:13:42Z kristl $
\name{predictionTest-class}
\Rdversion{1.1}
\docType{class}
\alias{predictionTest-class}
\title{Class "predictionTest"}
\description{
  A virtual class for all predictor test subclasses.  Currently
  subclasses \code{\linkS4class{PLSRTest}} and
  \code{\linkS4class{ridgeRegressionTest}} are defined.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "predictionTest" in the signature.
}
\author{Bjørn-Helge Mevik and Kristian Hovde Liland}
\seealso{Subclasses \code{\linkS4class{PLSRTest}} and
  \code{\linkS4class{ridgeRegressionTest}}.
}
\keyword{classes}
